requirements.txt
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
Django==2.1.7
django-haystack==2.8.1
django-redis==4.10.0
elasticsearch==2.4.1
fdfs-client-py==1.2.6
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.14.0
mutagen==1.42.0
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
PyMySQL==0.9.3
pytz==2019.1
qiniu==7.2.6
redis==3.2.1
requests==2.22.0
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
